+++ src/qt/KinectComunicator.hpp
@@ -17,7 +17,7 @@ public:
 
     void initialize();
 
-    virtual void processDepth(int time, const UINT16* buffer, int height, int width, USHORT minDepth, USHORT maxDepth);
+    virtual void processDepth(INT64 time, const UINT16* buffer, int height, int width, USHORT minDepth, USHORT maxDepth);
 
     private slots:
     void updateKinectManager();
